rename watch-machine to avoid tss route warnings
